[Saves] Uploading my current WIP concerning saves.
I added coSave for TES5, FO3 and TES4 and refactored some shared decoding functions in wbSaveInterface.
FO3 saves are only started and TES4 are not done at all. Neither can be called at the moment.

Started adding the option to open save or cosave files in View/Edit.
Added parameter G: for Game Saves Path.

Testing so far has been concentrated on verifying the rest still works as intended.

I also looked at forcing the date/time on skyrim plugins so the order in the editor is the same as in the runtime, but no luck so far.
Due to the issue with steam based mods, this may not be suitable for release, and is outside the scope of the tool anyway :)

I also confirmed the two Unknown in CTDA are unused (padding).


git-svn-id: http://skyrim-plugin-decoding-project.googlecode.com/svn@1794 bdfc8fb3-84cf-a68a-e902-ed0415cf5604