x64 compatibility fixes
Those are source changes to confirm that updating to x64 compatibiliy
does not break the 32 bit version.

Trying to compile this for Win64 will fail as it doesn't contain valid
subprojects or subproject references (Jedi, Treeview and mostly
TreeviewEdit)

The only computer I can generate a Win64 version on so far is my work
workstation. Cannot even repeat it at home (it refuses to add a Win64
platform so far :( )

Test binaries are included but don't expect this to be a regular thing
:) .