Added "Themes" and other numerous changes
Added "Themes" entry to the system menues of the Main form, File Select form and Options form. This can be used to choose from a number of different themes. The one called just "Windows" is the native windows theme. (And is the standard). The theme is saved in the settings and will be activated again on next start.

TwbFormID.ToString has now an aForDisplay:Boolean parameter. If true, ToString is allowed to format the string to make it prettier, otherwise it's guaranteed to keep returning an 8 character Hex string.

There is a flag wbPrettyFormIDs it's true for View and Edit modes by default, and can be deactivated with the -SimpleFormIDs command line paramter. When active, displayed FormIDs with have a space (or two for "light" modules) between FileID and ObjectID added.

xEdit should be "High DPI"-aware now.

The navigation treeview shows FormIDs in a monospace font.

Fix for issue #471: The code to verify that the saved window position was inside the current desktop was faulty.

VirtualTreeView has been updated to 7.x (current head from master in virtual tree view github repository).

The .dproj file has been completely replaced with a newly created one for Delphi 10.2 Tokyo.

version number increased to 3.2.3e